% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR_functions.R
\name{filterData}
\alias{filterData}
\title{Filtering based on on pValue and presentLimit.}
\usage{
filterData(data, pValue, presentLimit)
}
\arguments{
\item{lumi}{object where colnames(data)=sample IDs and rownames(data) = probe IDs}

\item{p-value}{limit pValue}
}
\value{
filtered lumi object where colnames(data)=sample IDs and rownames(data) = probe IDs
}
\description{
Filtering based on on pValue and presentLimit.
}
