% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR_functions.R
\name{mapToGenes}
\alias{mapToGenes}
\title{Map probes to genes.}
\usage{
mapToGenes(data)
}
\arguments{
\item{matrix}{where colnames(data)=sample IDs and rownames(data) = probe IDs}
}
\value{
matrix where colnames(data)=sample IDs and rownames(data) = gene IDs
}
\description{
Map probes to genes.
}
