% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR_functions.R
\name{normalizeData}
\alias{normalizeData}
\title{Normalization procedure from Vanesssa.}
\usage{
normalizeData(data.new)
}
\arguments{
\item{lumi}{object where colnames(data)=sample IDs and rownames(data) = probe IDs}
}
\value{
normalized matrix where colnames(data)=sample IDs and rownames(data) = probe IDs
}
\description{
Normalization procedure from Vanesssa.
}
