% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR_functions.R
\name{normalizeDataComBat}
\alias{normalizeDataComBat}
\title{Do ComBat normalization based on batch.}
\usage{
normalizeDataComBat(data.new, bTabName, sampleID, batchVar)
}
\arguments{
\item{data.new}{lumi object where colnames(data)=sample IDs and rownames(data) = probe IDs}

\item{bTabName}{name of the table, which is used for extracting batch variable}

\item{sampleID}{name of column in data.new dataframe, which represents sample ID number}

\item{batchVar}{batch variable, i.e. vector formed from dataset, which gives batch parameter for ComBat}
}
\value{
normalized matrix where colnames(data)=sample IDs and rownames(data) = probe IDs
}
\description{
Do ComBat normalization based on batch using sva library.
}
