% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR_functions.R
\name{normalizeDataVstQ}
\alias{normalizeDataVstQ}
\title{Log2 transformation using a variance stabilizing technique* and quantile normalization.}
\usage{
normalizeDataVstQ(data.new)
}
\arguments{
\item{data.new}{lumi object where colnames(data)=sample IDs and rownames(data) = probe IDs}
}
\value{
normalized matrix where colnames(data)=sample IDs and rownames(data) = probe IDs
}
\description{
* Lin SM, Du P, Huber W, Kibbe WA. Model-based variance-stabilizing transformation for 
Illumina microarray data. Nucleic Acids Res. 2008;36(2):e11.
}
\details{
Check also: http://journals.plos.org/plosone/article/file?id=10.1371/journal.pone.0156594&type=printable
}
