% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR_functions.R
\name{performBackgroundCorrection}
\alias{performBackgroundCorrection}
\title{Perform background correction and remove bad probes.}
\usage{
performBackgroundCorrection(data, exprs, negCtrl)
}
\arguments{
\item{lumi}{object where colnames(data)=sample IDs and rownames(data) = probe IDs}

\item{matrix}{where rownames(exprs)=sample IDs and colnames(exprs) = probe IDs}

\item{matrix}{where rownames(exprs)=sample IDs and colnames(data) = negative control IDs}
}
\value{
background corrected lumi object where colnames(data)=sample IDs and rownames(data) = gene IDs
}
\description{
Perform background correction and remove bad probes.
}
