% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR_functions.R
\name{performBackgroundCorrection}
\alias{performBackgroundCorrection}
\title{Perform background correction and remove bad probes.}
\usage{
performBackgroundCorrection(data, negCtrl)
}
\arguments{
\item{data}{lumi object with gene expression matrix exprs(data) where
colnames(exprs(data)) = sample IDs ( = labnr)
rownames(exprs(data)) = probe IDs}

\item{negCtrl}{matrix where
rownames(negCtrl) is a subset of colnames(exprs(data))
each column in negCtrl contains expression values for a negative control probe}
}
\value{
background-corrected lumi object where 
        colnames(exprs(data)) = sample IDs ( = labnr)
        rownames(exprs(data)) = probe IDs
}
\description{
Perform background correction and remove bad probes.
}
